Model Training and Evaluation
Investigate the performance of a number of machine learning procedures on this dataset. The machine learning procedures, like the decision tree classifier, SVM, and KNN approaches, are implemented, and the model is evaluated to predict if a new customer is likely to suffer from rising energy prices. In addition to this, various data preprocessing methods were also implemented. To design another system to predict the variation in the annual expenditure that a customer is going to have due to the increase in energy costs, different regression methods like the Random Forest approach, Linear regression, and decision tree regression methods are evaluated and compared.